# Small optimisation: move distance==0 comparison to outer loop